requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
asyncpg==0.21.0
attrs==20.1.0
certifi==2020.6.20
chardet==3.0.4
discord.py==1.4.1
idna==2.10
multidict==4.7.6
pytz==2020.1
yarl==1.5.1
